fix: resolve code-review findings across sbql-core, sbql-tui, and sbql-ffi (#12)
A high-effort multi-agent review of the workspace surfaced ~24 distinct
defects. This commit fixes them, most-severe first, with regression tests.

Data-safety (silent corruption / loss):
- Composite primary keys are carried in full through UPDATE/DELETE
  (CoreCommand pk is now Vec<(String,String)>), so a multi-column key
  addresses one row instead of every row sharing its first column.
- Row edits/deletes resolve their target table from the executed query
  (results.source_sql), not from unexecuted editor text.
- Editing a connection's target drops the stale pool + tunnel and resets
  session query state, so queries can't hit the old database.
- Paging is refused while edits/deletes are staged, and a discarded stage
  is always reported instead of vanishing silently.
- selected_row/col are clamped when a shorter page arrives.
- connections.toml is written atomically (temp + rename).
- Editing a connection no longer erases its SSH tunnel settings.

Security:
- SSH host-key verification (accept-new / TOFU via known_hosts); a changed
  key is refused.
- MySQL CSV import escapes backslashes, closing a string-literal breakout.

Correctness:
- LIMIT is detected as a token outside string literals (rate_limits no
  longer disables pagination and OOMs).
- Filter SQL is backend-aware (MySQL backticks + CAST AS CHAR, SQL Server
  brackets + NVARCHAR(MAX)); LIKE metacharacters are escaped with ESCAPE.
- Sort identifiers are quoted per backend; SQL Server ORDER BY carries
  OFFSET 0 ROWS so it is valid inside the pagination subquery.
- SQLite value suggestions declare ESCAPE '\'.
- JSON export escapes control characters; SQL export no longer maps empty
  strings to NULL.
- connection_string percent-encodes the user and database, not just the
  password.
- Unicode fixes: extract_schema_table_from_sql and completion prefix
  handling no longer byte-slice on char positions (café/ligatures panicked).
- SQLite PRAGMA statements quote identifiers via quote_ident.
- SSH keyring read errors are logged instead of silently swallowed.
- The SSH tunnel is torn down when the pool fails to connect.

sbql-ffi:
- Keychain support is now non-optional (removed the keyring feature;
  depend on sbql-core with features = ["keyring"]). This eliminates the
  --no-default-features build that silently dropped Keychain while save
  reported success, drops a redundant default-features flag, and lets the
  stale release-plz manifest comments be corrected.

+++ sbql-core/tests/mongodb_integration.rs
@@ -227,7 +227,7 @@ async fn test_mongodb_large_collection() {
 
     let coll = db.collection::<Document>("big");
     let docs: Vec<Document> = (0..150)
-        .map(|i| doc! { "idx": i as i32, "label": format!("row_{}", i) })
+        .map(|i| doc! { "idx": i, "label": format!("row_{}", i) })
         .collect();
     coll.insert_many(docs).await.unwrap();
